Revert "Fix: modificadas las variables aceleracion, velocidadX, velocidadY"
This reverts commit 72bd6a36897818f9e2d3cddc3ad8f2a270b6bc50.

+++ src/SpaceShip.java
@@ -2,9 +2,9 @@ public class SpaceShip {
 
     private String nombre;
     private String matriculaGalactica;
-    private int aceleracion;
-    private int velocidadX = 0;
-    private int velocidadY = 0;
+    private float aceleracion;
+    private float velocidadX = 0;
+    private float velocidadY = 0;
     private int coordenadaX = 0;
     private int coordenadaY = 0;
     private int direccionX = 0;
